chore: moving frontend routes folder to better suit MVVM architecture
