+++ front/src/lib/constants/components/Button.ts
@@ -1,3 +1,3 @@
-import type { ButtonRootSizes } from "../../../views/components/common/Button/Root/types";
+import type { ButtonRootSizes } from "../../components/common/Button/Root/types";
 
 export const DEFAULT_BUTTON_SIZE: ButtonRootSizes = 'DEFAULT';
