--- front/src/routes/+page.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-	import RankingTable from '../views/components/state/RankingTable.svelte';
-	import Selector from '../views/components/state/Selector.svelte';
-</script>
-
-<div class="person-page">
-	<Selector />
-	<RankingTable />
-</div>
-
-<style>
-</style>
