+++ front/src/views/routes/+page.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+	import RankingTable from '../components/state/RankingTable.svelte';
+	import Selector from '../components/state/Selector.svelte';
+</script>
+
+<div class="person-page">
+	<Selector />
+	<RankingTable />
+</div>
+
+<style>
+</style>
