+++ front/src/views/routes/types.ts
