+++ front/svelte.config.js
@@ -11,8 +11,12 @@ const config = {
 		// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
 		// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
 		// See https://svelte.dev/docs/kit/adapters for more information about adapters.
-		adapter: adapter()
-	}
+		adapter: adapter(),
+
+		files: {
+			routes: 'src/views/routes',
+		},
+	},
 };
 
 export default config;
